MDEV-26206 gap lock is not set if implicit lock exists
Move check if table is locked from implicit to explicit lock conversion
function to the higher level. This can increase performace by avoiding
per-record check and doing per-row check instead.

This is experimental commit, if testing is passed and perfomance
influence is good, the commit can be squashed with the previous one.